Render noExpander icon from icon map for leaf nodes
This is not the right answer and should not be merged: it works for FontAwesome where the .fa class is a good way to get an empty space. I don't know whether glyph icons has anything similar. And this doesn't solve the problem for non-glyph rendering. For reference, this patch requires this as part of a fontawesome config for the glyph extension: ```glyph:{map:{noExpander:"fa"}}```