Ensure tree item is focused *after* mouse click has had a chance to set focus.
Currently treeSetFocus will focus the first tree item if nothing else had focus.
However, a mouse click does not have a chance to set focus until after the
mouse is released. In order for the treeSetFocus to know that something
was focused with the mouse, we need to:
1) Set focus on a tree item immediately on a mousedown
2) Use a setTimeout(ensureFocus, 0) in treeSetFocus to ensure that
   it runs after the mousedown handler

This prevents the tree from scrolling back up when the user clicks in it.